Better folding of anonymous subs
Allowing any closing brace to end the fold of an anonymous sub makes it
too easy for things to get out of sync.  Change it to only consider a
closing brace at the beginning of a line preceded only by the same
amount of whitespace as exists at the beginning of the line on which the
anonymous sub started.  Only start a fold for anonymous subs where the
start of the fold is followed by nothing other than whitespace and a
possible comment, so that single-line subs don't start a fold for which
the end pattern wouldn't match.